requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
reportlab==4.0.8

